requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
einops>=0.7.0
bitsandbytes>=0.41.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
